# perf: Use `MDB_APPEND` when possible