--- a/dozer-types/src/borrow/mod.rs
+++ b/dozer-types/src/borrow/mod.rs
@@ -77,6 +77,7 @@ macro_rules! impl_borrow_for_clone_type {
 
 impl_borrow_for_clone_type!(u8, u32, u64, Record, IndexDefinition, SchemaWithIndex);
 
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Cow<'a, B: Borrow + 'a> {
     Borrowed(B::Borrowed<'a>),
     Owned(B),