Closes #1215 

# Testing method

With config file:

```yaml
connections:    
  - config : !LocalStorage
      details:
        path: data
      tables:
        - !Table
          name: trips
          prefix: /trips
          file_type: parquet
          extension: .parquet
    name: ny_taxi

endpoints:
  - name: trips
    path: /trips
    table_name: trips

cache_max_map_size: 10000000000
```

and data from `dozer-samples-connectors-local-storage` e2e test.

Comment https://github.com/getdozer/dozer/blob/main/dozer-api/src/cache_builder/log_reader.rs#L73.

```bash
dozer migrate
dozer app run
```

`time cargo run -- release dozer -- api run`

Hit Ctrl + C when counter reaches 3_000_000.

# Before

```text
cargo run --release -- api run  168.85s user 13.52s system 264% cpu 1:09.03 total
```

# After

```text
cargo run --release -- api run  64.95s user 0.91s system 464% cpu 14.188 total
```

5x speed up.